+++ lib/jazzy/highlighter.rb
@@ -6,5 +6,13 @@ module Jazzy
     def self.highlight(source, language)
       source && Rouge.highlight(source, language, 'html')
     end
+
+    def self.highlight_objc(source)
+      highlight(source, 'objc')
+    end
+
+    def self.highlight_swift(source)
+      highlight(source, 'swift')
+    end
   end
 end